Phew. Okay. I went ahead an made a lot of tiny passing-around-data changes which were needed to solve a larger bug than I anticipated.
Now, when a connection fails or a valid search isn't provided the result is None, while if a search yeilds no results the result is a blank list (or table). I updated documentation to reflect this.

For consistancy and readability, I also went forward and modfied the CrimeDetails system to work with smaller hunks of data. In the past, it was getting full 2d tables and then only using the first element, now it only fetches one row and handles the data in that a bit more directly.

I also fixed a few ugly try:catch: blocks.

Things are working on my thacker.